Move test driver into module
Starting to decouple test driver from Fluvio
